#31 - Upgraded to Spring Boot 1.4 GA.
Removed explicit configuration of JSR-310 converters as the Hibernate version Spring Boot 1.4 includes by default is able to persist JSR-310 types out of the box.

Removed explicit disabling of the thymeleaf caching as the DevTools module enables that out of the box.

Added Webjars locator library so that the template can use references to JavaScript libraries without referring to versions.