#54 - General overhaul.
Upgrade to Spring Boot 2.2 snapshots and Thin Launcher in 1.0.21. Tweaked application default configuration for quicker startup.

Upgraded to Bootstrap 4.3.1, and jQuery 3.3.1-2. Overhauled HTML template to work on Bootstrap 4. Made use of more Bootstrap features for error validation and styling. Tweaked messages to German.

Guestbook is now GuestbookRepository to avoid confusion. Switched to use CrudRepository in favor of declaring all methods manually.

GuestbookForm is now an immutable form DTO. A bit overcommented to explain what's going on an why it exists and is designed differently that the actual entity.